Added the different actions controlled with health/strength/agility stats, still need to set up the coming back to checkpoint for the sanctuary, but for that i need the deathscreen

+++ version_control/Assets/Scripts/Enemies/EnemyStates/Specific/Enemy1/Enemy1.cs
@@ -32,7 +32,7 @@ public class Enemy1 : Entity, IDamageable
         currentHealth = enemyData.maxHealth;
     }
 
-    public void Damage(float amount)
+    public void Damage(int amount)
     {
         currentHealth -= amount;
