+++ version_control/Assets/Scripts/Interfaces/IDamageable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void Damage(float amount);
+    void Damage(int amount);
 }
